Interim commit for creating a loaders module
