+++ biothings/hub/dataplugin/loaders/docker_dumper.py.tpl
